UCD-94 add audio support in core desktop
This first patch adds the rastersoft-gmail/pipewire-core PPA
with a more recent (and patched) pipewire version. The patched
version includes support for honoring the audio permissions.

It adds that PPA instead of the desktop-snappers one because
just uploading the .deb there without other changes here
would break the build. When this patch it merged, the .deb
will be uploaded to the desktop-snappers PPA and a new MR
removing rastersoft-gmail/pipewire-core will be sent.